Javascript variable name to include $ and _
Modification of the regex that select javascript arguments to include arguments containing $ and _.
In cases where a function had an argument with one of those, the function was not recognize (since the rex does't work in those cases).


That should cover 99% of the uses case, but note that a variable or an arguments in javascript can containt a lots of special character. See https://mathiasbynens.be/notes/javascript-identifiers for more informations.